Eliminate/workaround several compiler warnings.
--HG--
extra : convert_revision : svn%3A83215879-3e5a-4751-8c9d-778f44bb06a5/trunk%40729

+++ testing/thrusttest/random.h
@@ -17,7 +17,7 @@ template<typename T>
   {
       T value = 0;
         
-      for(int i = 0; i < sizeof(T); i++)
+      for(size_t i = 0; i < sizeof(T); i++)
           value ^= T(rand() & 0xff) << (8*i);
 
       return value;
